+++ src/constructors/css.js
@@ -3,6 +3,9 @@ import interleave from '../utils/interleave'
 import flatten from '../utils/flatten'
 import type { Interpolation, RuleSet } from '../types'
 
-export default (strings: Array<string>, ...interpolations: Array<Interpolation>): RuleSet => (
-  flatten(interleave(strings, interpolations))
-)
+export default (strings: Array<string>, ...interpolations: Array<Interpolation>): RuleSet => {
+  const rules = flatten(interleave(strings, interpolations))
+  // $FlowFixMe
+  rules.isStyledBlock = true
+  return rules
+}