+++ src/models/StyledComponent.js
@@ -15,7 +15,7 @@ import { CHANNEL } from './ThemeProvider'
 const escapeRegex = /[[\].#*$><+~=|^:(),"'`]/g
 const multiDashRegex = /--+/g
 
-export default (ComponentStyle: Function, constructWithOptions: Function) => {
+export default (ComponentStyle: Function, InlineStyle: Function, constructWithOptions: Function) => {
   /* We depend on components having unique IDs */
   const identifiers = {}
   const generateId = (_displayName: string) => {
@@ -42,6 +42,7 @@ export default (ComponentStyle: Function, constructWithOptions: Function) => {
     state = {
       theme: null,
       generatedClassName: '',
+      generatedStyles: null,
     }
 
     buildExecutionContext(theme: any, props: any) {
@@ -62,15 +63,33 @@ export default (ComponentStyle: Function, constructWithOptions: Function) => {
     }
 
     generateAndInjectStyles(theme: any, props: any) {
-      const { componentStyle, warnTooManyClasses } = this.constructor
+      const { componentStyle, styledComponentId, warnTooManyClasses } = this.constructor
       const executionContext = this.buildExecutionContext(theme, props)
-      const className = componentStyle.generateAndInjectStyles(executionContext)
+      const css = componentStyle.generateAndInjectStyles(executionContext)
 
       if (warnTooManyClasses !== undefined) {
-        warnTooManyClasses(className)
+        warnTooManyClasses(css)
       }
 
-      return className
+      const inlineRules = Object.keys(this.attrs).reduce((styles, name) => {
+        if (!Array.isArray(this.attrs[name])) {
+          return styles
+        }
+
+        if (!this.attrs[name].isStyledBlock) {
+          return styles
+        }
+
+        return styles.concat(this.attrs[name])
+      }, [])
+
+      const componentInlineStyle = new InlineStyle(inlineRules, styledComponentId)
+      const styles = componentInlineStyle.generateStyleObject(executionContext)
+
+      return {
+        generatedClassName: css,
+        generatedStyles: styles,
+      }
     }
 
     componentWillMount() {
@@ -87,16 +106,16 @@ export default (ComponentStyle: Function, constructWithOptions: Function) => {
           const { defaultProps } = this.constructor
           const isDefaultTheme = defaultProps && this.props.theme === defaultProps.theme
           const theme = this.props.theme && !isDefaultTheme ? this.props.theme : nextTheme
-          const generatedClassName = this.generateAndInjectStyles(theme, this.props)
-          this.setState({ theme, generatedClassName })
+          const styles = this.generateAndInjectStyles(theme, this.props)
+          this.setState({ theme, ...styles })
         })
       } else {
         const theme = this.props.theme || {}
-        const generatedClassName = this.generateAndInjectStyles(
+        const styles = this.generateAndInjectStyles(
           theme,
           this.props,
         )
-        this.setState({ theme, generatedClassName })
+        this.setState({ theme, ...styles })
       }
     }
 
@@ -107,9 +126,9 @@ export default (ComponentStyle: Function, constructWithOptions: Function) => {
         const { defaultProps } = this.constructor
         const isDefaultTheme = defaultProps && nextProps.theme === defaultProps.theme
         const theme = nextProps.theme && !isDefaultTheme ? nextProps.theme : oldState.theme
-        const generatedClassName = this.generateAndInjectStyles(theme, nextProps)
+        const styles = this.generateAndInjectStyles(theme, nextProps)
 
-        return { theme, generatedClassName }
+        return { theme, ...styles }
       })
     }
 
@@ -121,7 +140,7 @@ export default (ComponentStyle: Function, constructWithOptions: Function) => {
 
     render() {
       const { children, innerRef } = this.props
-      const { generatedClassName } = this.state
+      const { generatedClassName, generatedStyles } = this.state
       const { styledComponentId, target } = this.constructor
 
       const isTargetTag = isTag(target)
@@ -151,6 +170,7 @@ export default (ComponentStyle: Function, constructWithOptions: Function) => {
           return acc
         }, {
           ...this.attrs,
+          style: generatedStyles || this.attrs.style,
           className,
           ref: innerRef,
           innerRef: undefined,
@@ -160,11 +180,9 @@ export default (ComponentStyle: Function, constructWithOptions: Function) => {
     }
   }
 
-  const createStyledComponent = (
-    target: Target,
-    options: Object,
-    rules: RuleSet,
-  ) => {
+  const createStyledComponent = (target: Target,
+                                 options: Object,
+                                 rules: RuleSet,) => {
     const {
       displayName = isTag(target) ? `styled.${target}` : `Styled(${target.displayName})`,
       componentId = generateId(options.displayName),