+++ src/no-parser/index.js
@@ -12,6 +12,7 @@ import styleSheet from '../models/StyleSheet'
 /* Import singleton constructors */
 import _StyledComponent from '../models/StyledComponent'
 import _ComponentStyle from '../models/ComponentStyle'
+import _InlineStyle from '../models/InlineStyle'
 import _GlobalStyle from '../models/GlobalStyle'
 import _styled from '../constructors/styled'
 import _keyframes from '../constructors/keyframes'
@@ -28,7 +29,7 @@ import withTheme from '../hoc/withTheme'
 const GlobalStyle = _GlobalStyle(flatten, stringifyRules)
 const ComponentStyle = _ComponentStyle(generateAlphabeticName, flatten, stringifyRules)
 const constructWithOptions = _constructWithOptions(css)
-const StyledComponent = _StyledComponent(ComponentStyle, constructWithOptions)
+const StyledComponent = _StyledComponent(ComponentStyle, _InlineStyle(generateAlphabeticName), constructWithOptions)
 
 /* Instantiate exported singletons */
 const keyframes = _keyframes(generateAlphabeticName, GlobalStyle, css)